test(DeviceAssociation): Split tests is dedicated files
